refactor: replace default manuscripts with elife's reference manuscripts
This is just a stop-gap solution until the integration with the new manuscript
backend is completed.
I've added the 'reference-manuscripts' repostitory as a submodule, and updated
the 'build' target to run the substance-bundler to pack the manuscripts into a
vfs file. As such, articles can now be opened in the editor by setting the
'archive' query parameter to the DOI of the desired article, e.g. 00001.