fix: only bundle manuscript xml files in vfs.js
I noticed that clients where taking a fair while to load the editor, and the reason being
that the vfs.js file was around the 67MB mark. I've updated the bundler to only package
xml files for now which means that images/figures won't load in the editor for now.
This will be resolved properly under https://github.com/libero/producer/issues/212.