fix: images not displayed in articles
Images are fetched on demand via the web-server rather than read from the VFS
file. As such, I now pack the manuscripts with the container so that the image
files can be serverd to clients by the web-server.
This is just a temporary solution until the real backend is created, but for
now it will do.